chore(openshift): rename container.ts -> deploy.ts
