+++ core/src/plugins/openshift/deploy.ts
