+++ core/src/plugins/openshift/openshift.ts
@@ -9,7 +9,7 @@
 import { ConfigureProviderParams } from "../../plugin/handlers/Provider/configureProvider"
 import { createGardenPlugin } from "../../plugin/plugin"
 import { OpenShiftConfig, configSchema } from "./config"
-import { openshiftContainerDeployExtension } from "./container"
+import { openshiftContainerDeployExtension } from "./deploy"
 
 export async function configureProvider({ config }: ConfigureProviderParams<OpenShiftConfig>) {
   return { config }
